fix(security): deep audit — harden remaining modules across all surfaces
ADMIN DASHBOARD (CRITICAL):
- Middleware now validates JWT exp claim instead of only checking cookie presence
- Reduce admin cookie max-age from 30 days to 8 hours (standard admin session)
- Switch token persistence from localStorage to sessionStorage (XSS mitigation)
- Gate all console.log/error calls behind NODE_ENV === development
- Remove dev_otp exposure from sendOtp response type
- Fix sendRideInvoice to use shared request() helper (was bypassing 401 handling)
- Add 429 rate-limit feedback on login page

BACKEND ROUTES (HIGH):
- Replace raw dict Body(...) with Pydantic UpdateDriverProfileRequest on PUT /drivers/me
- Eliminate global stripe.api_key mutation; use per-call api_key= pattern (async safety)
- Remove mock payment intent fallback; raise 503 when Stripe not configured
- Reject unverified Stripe webhooks (was silently accepting with verified=False)
- Migrate promotions money fields from float to Decimal (CQ-009 compliance)
- Migrate disputes money fields from float to Decimal
- Replace raw request.json() with RegisterTokenRequest Pydantic model in notifications

BACKEND CORE (HIGH):
- Fix division-by-zero in point_in_polygon (horizontal edge guard)
- Add WebSocket message rate limiting (30 msg/sec) and size cap (64KB)
- Gate diagnostic loggers (goonline.debug, spinr.diag) behind DEBUG env var

DRIVER APP (MEDIUM):
- Replace ActiveRide any types with typed RideInfo/RiderInfo/VehicleTypeInfo interfaces

SHARED PACKAGE (MEDIUM):
- Use fetchWithTimeout consistently across put/patch/delete (was raw fetch)
- Gate all 14+ console.log statements in authStore behind __DEV__

FRONTEND WEB (LOWER):
- Switch web token storage from localStorage to sessionStorage

+++ backend/routes/disputes.py
@@ -5,6 +5,7 @@ disputes.py – Payment dispute/refund request endpoints for Spinr.
 import logging
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import Any, Dict, Optional
 
 from fastapi import APIRouter, Depends, HTTPException
@@ -26,12 +27,12 @@ class CreateDisputeRequest(BaseModel):
     ride_id: str
     reason: str  # overcharged | wrong_route | driver_issue | payment_error | other
     description: str
-    requested_amount: Optional[float] = None  # If blank, full refund
+    requested_amount: Optional[Decimal] = None  # If blank, full refund
 
 
 class ResolveDisputeRequest(BaseModel):
     resolution: str  # approved | partial_refund | rejected
-    refund_amount: Optional[float] = None
+    refund_amount: Optional[Decimal] = None
     admin_note: Optional[str] = None
